# docs/swagger.yaml
basePath: /
definitions:
...
  requestmodels.Property:
    properties:
      Apartment:
        type: string
      Balcony:
        type: boolean
      City:
        type: string
      Country:
        type: string
      Garage:
        type: boolean
      Garden:
        type: boolean
      HeatType:
        type: integer
      Housenumber:
        type: string
      Landlord:
        type: string
      Rooms:
        type: integer
      Size:
        type: integer
      Status:
        type: integer
      Street:
        type: string
      Tenant:
        type: string
      Type:
        type: integer
    type: object
  requestmodels.RentProfile:
    properties:
      ContractDue:
        type: string
      Deposit:
        type: number
      HeatCosts:
        type: number
      Minimum:
        type: integer
      Rent:
        type: number
      RentalStart:
        type: string
    type: object
  requestmodels.SearchProfile:
    properties:
      Budget:
        type: number
      City:
        type: string
      Country:
        type: string
      Radius:
        type: number
      Rooms:
        type: integer
      Size:
        type: integer
      Street:
        type: string
      Type:
        type: integer
      Zipcode:
        type: string
    type: object
...
  responsemodels.Property:
    properties:
      Apartment:
        type: string
      Balcony:
        type: boolean
      City:
        type: string
      Country:
        type: string
      Garage:
        type: boolean
      Garden:
        type: boolean
      HeatType:
        type: integer
      Housenumber:
        type: string
      Id:
        type: string
      Landlord:
        type: string
      Rooms:
        type: integer
      Size:
        type: integer
      Status:
        type: integer
      Street:
        type: string
      Tenant:
        type: string
      Type:
        type: integer
    type: object
  responsemodels.RentProfile:
    properties:
      ContractDue:
        type: string
      Deposit:
        type: number
      HeatCosts:
        type: number
      Id:
        type: string
      Minimum:
        type: integer
      Rent:
        type: number
      RentalStart:
        type: string
    type: object
  responsemodels.SearchProfile:
    properties:
      Budget:
        type: number
      City:
        type: string
      Country:
        type: string
      Radius:
        type: number
      Rooms:
        type: integer
      Size:
        type: integer
      Street:
        type: string
      Type:
        type: integer
      UserId:
        type: string
      Zipcode:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
...
  /property:
    get:
      consumes:
      - application/json
      description: Get all properties
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responsemodels.Property'
            type: array
      summary: Get all properties
      tags:
      - Property
    post:
      consumes:
      - application/json
      description: Create a new property
      parameters:
      - description: Property
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/requestmodels.Property'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Property'
      summary: Create a new property
      tags:
      - Property
  /property/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete a property
      tags:
      - Property
    get:
      consumes:
      - application/json
      description: Get a property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Property'
      summary: Get a property
      tags:
      - Property
    put:
      consumes:
      - application/json
      description: Update a property
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: string
      - description: Property
        in: body
        name: property
        required: true
        schema:
          $ref: '#/definitions/requestmodels.Property'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Property'
      summary: Update a property
      tags:
      - Property
...
  /property/landlord/{id}/{landlordId}:
    put:
      consumes:
      - application/json
      description: Update a property landlord
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: string
      - description: Landlord ID
        in: path
        name: landlordId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update a property landlord
      tags:
      - Property
...
  /property/rentprofile/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a rent profile
      operationId: delete-rent-profile
      parameters:
      - description: Rent Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete a rent profile
    get:
      consumes:
      - application/json
      description: Get a rent profile
      operationId: get-rent-profile
      parameters:
      - description: Rent Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/responsemodels.RentProfile'
      summary: Get a rent profile
    post:
      consumes:
      - application/json
      description: Create a rent profile
      operationId: create-rent-profile
      parameters:
      - description: Rent Profile
        in: body
        name: rentProfile
        required: true
        schema:
          $ref: '#/definitions/requestmodels.RentProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.RentProfile'
      summary: Create a rent profile
    put:
      consumes:
      - application/json
      description: Update a rent profile
      operationId: update-rent-profile
      parameters:
      - description: Rent Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Rent Profile
        in: body
        name: rentProfile
        required: true
        schema:
          $ref: '#/definitions/requestmodels.RentProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.RentProfile'
      summary: Update a rent profile
  /property/tenant/{id}/{tenantId}:
    put:
      consumes:
      - application/json
      description: Update a property tenant
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update a property tenant
      tags:
      - Property
...
  /user/searchprofile/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user's search profile
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete a user's search profile
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get a user's search profile
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/responsemodels.SearchProfile'
      summary: Get a user's search profile
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create a user's search profile
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Search Profile
        in: body
        name: searchProfile
        required: true
        schema:
          $ref: '#/definitions/requestmodels.SearchProfile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responsemodels.SearchProfile'
      summary: Create a user's search profile
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update a user's search profile
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Search Profile
        in: body
        name: searchProfile
        required: true
        schema:
          $ref: '#/definitions/requestmodels.SearchProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.SearchProfile'
      summary: Update a user's search profile
      tags:
      - user
...
schemes:
- http
swagger: "2.0"